Spendly-main/backend/requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
python-multipart
pydantic
groq